rename main folder
