+++ project/zadanie_1.cpp
